+++ src/Parameters.ts
@@ -1,5 +1,4 @@
 import { Logger } from "./Logger";
-import { AuthCredentials } from "./pac/auth/authParameters";
 
 export interface LoggerParameters
 {
@@ -15,18 +14,4 @@ export interface RunnerParameters extends LoggerParameters, TelemetryParameters
 {
   workingDir: string;
   runnersDir: string;
-}
-
-export interface WhoAmIParameters
-{
-  credentials: AuthCredentials;
-  environmentUrl: string;  
-}
-
-export interface ExportSolutionParameters
-{
-  name: string; 
-  path: string;
-  credentials: AuthCredentials;
-  environmentUrl: string;  
 }
